+++ src/services/userActivity.js
@@ -8,7 +8,7 @@ import { DataContext, UserActivityVersionKey } from './dataContext.js'
 import { fetchByRailContentIds, fetchParentChildRelationshipsFor } from './sanity'
 import { getMonday, getWeekNumber, isSameDate, isNextDay } from './dateUtils.js'
 import { globalConfig } from './config'
-import { addAwardTemplateToContent, getFormattedType } from '../contentTypeConfig'
+import { postProcessBadge, getFormattedType } from '../contentTypeConfig'
 import dayjs from 'dayjs'
 import { addContextToContent } from './contentAggregator.js'
 import { db, Q } from './sync'
@@ -505,7 +505,7 @@ export async function getRecentActivity({ page = 1, limit = 5, tabName = null } 
       addNextLesson: true,
     }
   )
-  contents = addAwardTemplateToContent(contents)
+  contents = postProcessBadge(contents)
 
   contents = await mapContentsThatWereLastProgressedFromMethod(contents)
 
@@ -790,7 +790,7 @@ async function formatPracticeMeta(practices = []) {
       addNextLesson: true,
     }
   )
-  contents = addAwardTemplateToContent(contents)
+  contents = postProcessBadge(contents)
 
   contents = await mapContentsThatWereLastProgressedFromMethod(contents)